CloudKitToObject create() uses the passed in type to init(). Now returns an object of the subclass type not just type Object.
Realm purgeDeletedObjects() filters on 'isDeleted'. This would have been a bug as it deleted all objects of the type beforehand irrespective of the 'isDeleted' flag.

NewSyncEngine is not generic and is instantiated NewSyncEngine(objectType: MyObject.self) as opposed to SyncEngine<MyObject>(). This paves the way for Syncing multiple object types and preserves existing API for now.