Pass completion block in SyncEngine.pull
As `DatabaseManager.fetchChangesInDatabase` supports passing in a completion block which is run once each zone is fetched lets pass that through as an optional void completion block.

This is a no-op for current users as the block is optional and defaults to nil but it means that we can now pass in a block which is only run once the pull is complete.

Whilst we could just simply observe the Realm itself for these changes there is no way to wait for a signal that all of a given batch (cloud pull) is done.