FIX: stripped emojis from CKRecord name to avoid crashing with invalid primary key
Bug: the app would crash if a primary key was used that had an emoji in it.
This was an issue for my app because the app accidentally used an emoji in
a primary key if the user chose to include an emoji in the title of something.
Fix: removed all emojis from the primary key if they existed. This was enough
to allow me to remove the error when saving primary keys and allow existing
users to continue to use their iCloud without resetting their cached data.